+++ db-client-java/src/main/java/com/eventstore/dbclient/ConnectionService.java
@@ -13,7 +13,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 /**
  * Main gRPC connection management service.
  */
-class ConnectionService implements Runnable, MsgHandler {
+class ConnectionService implements Runnable {
     private final Logger logger = LoggerFactory.getLogger(ConnectionService.class);
     private final GrpcClient client;
     private final AtomicBoolean closed;
@@ -99,7 +99,6 @@ class ConnectionService implements Runnable, MsgHandler {
         this.forceExit(null);
     }
 
-    @Override
     public void createChannel(UUID previousId, InetSocketAddress candidate) {
         if (this.closed.get()) {
             logger.warn("Channel creation request ignored, the connection to endpoint [{}] is already closed", this.connection.getLastConnectedEndpoint());
@@ -128,6 +127,13 @@ class ConnectionService implements Runnable, MsgHandler {
                 try {
                     // TODO - Should we consider a discovery timeout?
                     this.discovery.run(this.connection).get();
+
+                    if (this.loadServerFeatures()) {
+                        this.channelId = UUID.randomUUID();
+                        this.connection.confirmChannel();
+                        logger.info("Connection to endpoint [{}] created successfully", this.connection.getLastConnectedEndpoint());
+                        break;
+                    }
                 } catch (InterruptedException e) {
                     forceExit(e);
                 } catch (ExecutionException e) {
@@ -136,15 +142,7 @@ class ConnectionService implements Runnable, MsgHandler {
                     // that has failed. It's possible that node might still be the best candidate if it manages to
                     // recover in the meantime.
                     this.connection.clear();
-                    continue;
                 }
-            }
-
-            if (this.loadServerFeatures()) {
-                this.channelId = UUID.randomUUID();
-                this.connection.confirmChannel();
-                logger.info("Connection to endpoint [{}] created successfully", this.connection.getLastConnectedEndpoint());
-                break;
             }
 
             // In case a candidate was provided, but we failed to connect to it.
@@ -154,7 +152,6 @@ class ConnectionService implements Runnable, MsgHandler {
         }
     }
 
-    @Override
     public void process(RunWorkItem args) {
         if (this.closed.get()) {
             logger.warn("Receive a command request but the connection to endpoint [{}] is already closed", this.connection.getLastConnectedEndpoint());
@@ -183,7 +180,6 @@ class ConnectionService implements Runnable, MsgHandler {
         args.getItem().accept(workArgs, null);
     }
 
-    @Override
     public void shutdown(Shutdown args) {
         if (this.closed.get()) {
             args.complete();